# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenAccounting API
  version: "1.0.0"
  description: |
    REST API for the OpenAccounting accounting platform.

    ## Authentication
    All `/api/v1/...` endpoints require a Bearer token:
    ```
    Authorization: Bearer oa_live_…
    ```

    ## Idempotency
    `POST` endpoints accept an `Idempotency-Key` header. Replays
    within 24 hours return the original response without re-executing
    the mutation. Keys are scoped per token + ledger.

    ## Pagination
    List endpoints return `{"data": [...]}` with a `Link: <…>; rel="next"`
    header when more results are available. Pass `?limit=` (default 50,
    max 200) and `?cursor=` from the previous `Link` header.

    ## Errors
    All errors return RFC 7807 problem-details JSON
    (`Content-Type: application/problem+json`).

    ## Incoming Events
    External systems can push signed events to
    `POST /api/events/{ledger_id}` using the ledger's
    `incoming_events_secret` (HMAC-SHA256 in the
    `X-OA-Event-Signature: sha256=<hex>` header).

    ## Automation Rules
    Ledger owners create rules via the web UI at
    `/ledgers/{id}/automations`. Rules map an allowlisted trigger
    (event type) + conditions → an action (webhook / email /
    categorise transaction) and execute on the jobs queue.
servers:
  - url: /
    description: Same origin
paths:
...
  /ledgers/{ledger_id}/accounts:
    get:
      operationId: listAccounts
      summary: List accounts in a ledger
      tags: [Accounts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountList"
    post:
      operationId: createAccount
      summary: Create an account
      tags: [Accounts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /ledgers/{ledger_id}/accounts/{id}:
    get:
      operationId: getAccount
      summary: Get a single account
      tags: [Accounts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/AccountIdPath"
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAccount
      summary: Update account name, description, or archive status
      tags: [Accounts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/AccountIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ledgers/{ledger_id}/budgets:
    get:
      operationId: listBudgets
      summary: List budgets in a ledger
      tags: [Budgets]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/PaginationParams"
      responses:
        "200":
          description: Budget list with vs-actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetList"
    post:
      operationId: createBudget
      summary: Create a budget (idempotent)
      tags: [Budgets]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
  /ledgers/{ledger_id}/budgets/{id}:
    get:
      operationId: getBudget
      summary: Get a single budget
      tags: [Budgets]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/BudgetIdPath"
      responses:
        "200":
          description: Budget details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteBudget
      summary: Delete a budget
      tags: [Budgets]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/BudgetIdPath"
      responses:
        "204":
          description: Budget deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    LedgerIdPath:
      name: ledger_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AccountIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BudgetIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Client-generated key for idempotent POSTs (24 h window)
    PaginationParams:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: cursor
        in: query
        required: false
        schema:
          type: string
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
          nullable: true
        type:
          type: string
        subtype:
          type: string
        currency:
          type: string
        is_archived:
          type: boolean
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
    CreateAccountBody:
      type: object
      required: [name, account_type, account_subtype]
      properties:
        name:
          type: string
        code:
          type: string
        account_type:
          type: string
        account_subtype:
          type: string
        description:
          type: string
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_archived:
          type: boolean
...
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account:
          type: string
        period:
          type: string
        amount:
          type: number
        alert_threshold:
          type: number
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    BudgetList:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
    CreateBudgetBody:
      type: object
      required: [account_id, period, amount, start_date, end_date]
      properties:
        account_id:
          type: string
          format: uuid
        period:
          type: string
          enum: [monthly, quarterly, yearly]
        amount:
          type: number
        alert_threshold:
          type: number
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
...